[+] ALF-911. Необходимо отображать оргструктуру в виде графа
git-svn-id: http://svn.datateh.ru/alfrescoecm/Business%20Platform/trunk@2660 697f8a5b-bb4b-4f4c-ab90-ff1ad96ece97

+++ lecm-orgstructure/repo/src/main/java/ru/it/lecm/orgstructure/scripts/diagram/OrgstructureUnit.java
@@ -70,7 +70,7 @@ public class OrgstructureUnit {
 
         fm = bi.getGraphics().getFontMetrics(fontBoss);
         if (boss != null && fm.stringWidth(boss) > width) {
-            width = (int) (fm.stringWidth(boss));
+            width = fm.stringWidth(boss);
         }
 
         fm = bi.getGraphics().getFontMetrics(font);
